More robust request error handling
Previously the APIs popped out of the loop whenever a server was
"contacted" succesfully (i.e., there we no protocol errors). So in case
a server was contacted but couldn't release an AC (e.g., the
user is not a member of the VO or is suspended) the loop would be
interrupted without trying a connection to another server.

Now all servers are tried until either an AC is received or requests to all
known replicas for a given VO alias have failed.