Add ability to set openssl validator hash function
The CANL Openssl certificate validator can be configured,
at creation time, to use the MD5 or SHA1 function to resolve
trust anchors. MD5 was used in openssl pre 1.x, SHA1 is used
since 1.x. Typically CA trust anchors are packaged in a way
that both hash functions would work, but we should not hide the
possibility of configuring which is the desired behaviour in the
CertificateValidatorBuilder.

Now CertificateValidatorBuilder allows to choose which has function
should be used, via the opensslHashFunction method.

The default hash function is MD5.

Issue: https://issues.infn.it/jira/browse/VOMS-703